feature: use redis for faster scraping
