+++ scraper/package.json
@@ -5,7 +5,8 @@
   "type": "module",
   "dependencies": {
     "@types/node": "^20.14.11",
-    "big-json": "^3.2.0"
+    "big-json": "^3.2.0",
+    "redis": "^4.7.0"
   },
   "devDependencies": {
     "@eslint/js": "^9.7.0",
@@ -17,7 +18,7 @@
     "globals": "^15.8.0",
     "luxon": "^3.4.4",
     "readline": "^1.3.0",
-    "tsx": "^4.16.2",
+    "tsx": "^4.16.3",
     "typescript": "^5.5.4"
   }
 }
